DDED: support for flight dataset
The flight data set had some serious issues with the date in various formats and invalid characters. In order to fix this a ruby script was used to remove invalid characters (that postgresql could not handle) and use DateTime.parse to generically covert the date string to a unix timestamp.

Refactored a use of Java Array to ArrayList since ArrayLists were used more commonly and allowed for easier use.

Moved test of a specific database into a single method within the main test class. This just allowed for the two database tests to be kept more or less isolated.

Added the database schema for the clean_flight dataset. As well as some general notes about the MFDs and usage.

The database interface was modified to accommodate bigint and serial. Also, added a update methods function to update the values changed during the cleaning of one MFD.

Finally, the repair method was modified to accomidate NULL fields. However this accomidation is not complete, since the algorithm now makes a false assumption that ever row will have a corresponding CorePattern which shares the X attributes. In the event that a set of tuples that share a set of X attributes these tuples will not have a CorePattern in the case that the Y attribute value for every tuple in the set is null.