ADDED: normalized distance calculation
In order to allow for a responable comparision between the LHS and RHS the distance function is normalized. Currently the normalization is done by dividing th distance by the maximum distance for the particular column.

For the LHS the column may consist of multiple columns and therefore the total normalized distance is calculated by taking the summation of the normalized distance for each attribute value divided by the number of attributes.

Finally, a basic repair class was created since the anonymous definition of the same class seemed redundent.